requirements.txt
jupyter
pandas
matplotlib
seaborn
mysql-connector-python